+++ run_optimizer.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python3
+"""
+Entry point script for Madrid Route Optimizer.
+Runs the modular route optimization application.
+"""
+
+from route_optimizer import main
+
+if __name__ == "__main__":
+    main()
